+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added `swipeActionsViewStyle` to `ListEnvironment`. This allows a `SwipeActionsView.Style` to be set on the list environment when customizing the appearance of the swipe action view.
 - Added the ability to configure leading swipe actions on `Item`s.
 - Added `containerCornerRadius`, `buttonSizing`, `minWidth`, and `maxWidthRatio` to `DefaultSwipeActionsView.Style` for additional swipe action style customization.
+- Added `SwipeActionsView.Style.leadingContainerInsets` to specify container insets for the leading swipe action container.
 
 ### Removed
 
@@ -22,6 +23,7 @@
 - The type of the `ItemContent.swipeActionsStyle` protocol requirement is now `SwipeActionsView.Style?` (previously `SwipeActionsView.Style`). When an item returns `nil` for this property the style set on the list environment will be used instead.
 - Renamed `Item.swipeActions` to `Item.leadingSwipeActions`.
 - Renamed `DefaultItemProperties.swipeActions` to `leadingSwipeActions`.
+- Renamed `SwipeActionsView.Style.containerInsets` to `SwipeActionsView.Style.trailingContainerInsets` and changed the type to `NSDirectionalEdgeInsets`.
 
 ### Misc